Add xcodeproj - carthage support
