+++ .gitignore
@@ -1,11 +1,18 @@
-.DS_Store
-sourcekitten-output.json
-docs/
-/.build
-/Packages
-/*.xcodeproj
-**/xcuserdata
-**/xcshareddata
+## Build generated
+build/
+DerivedData/
+
+## Various settings
+*.pbxuser
+!default.pbxuser
+*.mode1v3
+!default.mode1v3
+*.mode2v3
+!default.mode2v3
+*.perspectivev3
+!default.perspectivev3
+xcuserdata/
+
+# Dependencies
 Pods/
 Carthage/
-Examples/**/Podfile.lock
