+++ SafeCollection/Podfile
@@ -1,0 +1,5 @@
+platform :ios, '8.0'
+
+target 'SafeCollectionTests' do
+  pod 'SafeCollection', :path => '../'
+end
