+++ Sources/SafeCollection.h
@@ -1,0 +1,4 @@
+#import <UIKit/UIKit.h>
+
+FOUNDATION_EXPORT double SafeCollectionVersionNumber;
+FOUNDATION_EXPORT const unsigned char SafeCollectionVersionString[];
