+++ Sources/SafeCollection.swift
